requirements.txt
anyio==3.7.1
certifi==2023.5.7
exceptiongroup==1.1.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
pip==23.1.2
python-telegram-bot==20.4
setuptools==68.0.0
sniffio==1.3.0
typing_extensions==4.7.1
wheel==0.40.0
